requirement.txt
streamlit
torch
torchvision
transformers==4.45.2
Pillow
matplotlib
IPython
opencv-python
flask